# schema-stub/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "IdentityChain Agent REST API"
  version: "0.1.6"
  title: "IdentityChain API"
schemes:
- "http"
consumes:
- "application/json"
produces:
- "application/json"
security:
- Bearer: []
paths:
...
  /api/connection/{myDid}:
    get:
      tags:
      - "connection"
      summary: "Retrieve a (pending) connection"
      description: "Retrieve a (pending) connection"
      operationId: "apiConnectionMyDidGET"
      parameters:
      - name: "myDid"
        in: "path"
        description: "A unique did used in a pairwise"
        required: true
        type: "string"
      - name: "wallet"
        in: "header"
        description: "(Optional) A unique id value (or \"default\") identifying a\
          \ wallet"
        required: false
        type: "string"
      responses:
        200:
          description: ""
          schema:
            $ref: "#/definitions/inline_response_200"
        404:
          description: "no matching connection found"
      x-swagger-router-controller: "Connection"
  /api/connectionoffer:
    get:
      tags:
      - "connection"
      summary: "List all my connection offers"
      description: "List all my connection offers"
      operationId: "apiConnectionofferGET"
      parameters:
      - name: "wallet"
        in: "header"
        description: "(Optional) A unique id value (or \"default\") identifying a\
          \ wallet"
        required: false
        type: "string"
      responses:
        200:
          description: ""
      x-swagger-router-controller: "Connection"
    post:
      tags:
      - "connection"
      summary: "Create a new connection offer"
      description: "Create a new connection offer"
      operationId: "apiConnectionofferPOST"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/connectionoffer_post"
      responses:
        201:
          description: ""
      x-swagger-router-controller: "Connection"
  /api/connectionoffer/{connectionOfferId}:
    get:
      tags:
      - "connection"
      summary: "Retrieve a connection offer by _id"
      description: "Retrieve a connection offer by _id"
      operationId: "apiConnectionofferConnectionOfferIdGET"
      parameters:
      - name: "wallet"
        in: "header"
        description: "(Optional) A unique id value (or \"default\") identifying a\
          \ wallet"
        required: false
        type: "string"
      - name: "connectionOfferId"
        in: "path"
        description: "connection offer _id"
        required: true
        type: "string"
      responses:
        200:
          description: ""
      x-swagger-router-controller: "Connection"
    delete:
      tags:
      - "connection"
      summary: "Delete a connection offer by _id"
      description: "Delete a connection offer by _id"
      operationId: "apiConnectionofferConnectionOfferIdDELETE"
      parameters:
      - name: "wallet"
        in: "header"
        description: "(Optional) A unique id value (or \"default\") identifying a\
          \ wallet"
        required: false
        type: "string"
      - name: "connectionOfferId"
        in: "path"
        description: "connection offer _id"
        required: true
        type: "string"
      responses:
        204:
          description: ""
      x-swagger-router-controller: "Connection"
  /api/connectionrequest:
    get:
      tags:
      - "connection"
      summary: "List all my connection requests"
      description: "List all my connection requests"
      operationId: "apiConnectionrequestGET"
      parameters:
      - name: "wallet"
        in: "header"
        description: "(Optional) A unique id value (or \"default\") identifying a\
          \ wallet"
        required: false
        type: "string"
      responses:
        200:
          description: ""
      x-swagger-router-controller: "Connection"
    post:
      tags:
      - "connection"
      summary: "Accept a connection offer and/or send a connection request"
      description: "Accept a connection offer and/or send a connection request"
      operationId: "apiConnectionrequestPOST"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/connectionrequest_post"
      responses:
        200:
          description: "Success"
      x-swagger-router-controller: "Connection"
  /api/connectionrequest/{connectionRequestId}:
    get:
      tags:
      - "connection"
      summary: "Retrieve a connection request by _id"
      description: "Retrieve a connection request by _id"
      operationId: "apiConnectionrequestConnectionRequestIdGET"
      parameters:
      - name: "connectionRequestId"
        in: "path"
        description: "connection request _id"
        required: true
        type: "string"
      - name: "wallet"
        in: "header"
        description: "(Optional) A unique id value (or \"default\") identifying a\
          \ wallet"
        required: false
        type: "string"
      responses:
        200:
          description: ""
      x-swagger-router-controller: "Connection"
    delete:
      tags:
      - "connection"
      summary: "Delete a connection request by _id"
      description: "Delete a connection request by _id"
      operationId: "apiConnectionrequestConnectionRequestIdDELETE"
      parameters:
      - name: "connectionRequestId"
        in: "path"
        description: "connection request _id"
        required: true
        type: "string"
      - name: "wallet"
        in: "header"
        description: "(Optional) A unique id value (or \"default\") identifying a\
          \ wallet"
        required: false
        type: "string"
      responses:
        204:
          description: ""
      x-swagger-router-controller: "Connection"
  /api/connectionresponse:
    post:
      tags:
      - "connection"
      summary: "Accept a connection request and send a connection response"
      description: "Accept a connection request and send a connection response"
      operationId: "apiConnectionresponsePOST"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/connectionresponse_post"
      responses:
        200:
          description: "Success"
      x-swagger-router-controller: "Connection"
...
  /api/credentialdef:
    get:
      tags:
      - "credential"
      summary: "List credential definitions of wallet"
      description: "List credential definitions of wallet"
      operationId: "apiCredentialdefGET"
      parameters:
      - name: "wallet"
        in: "header"
        description: "(Optional) A unique id value (or \"default\") identifying a\
          \ wallet"
        required: false
        type: "string"
      responses:
        200:
          description: ""
      x-swagger-router-controller: "Credential"
    post:
      tags:
      - "credential"
      summary: "Create a credential definition"
      description: "Create a credential definition"
      operationId: "apiCredentialdefPOST"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/credentialdef_post"
      responses:
        201:
          description: ""
      x-swagger-router-controller: "Credential"
  /api/credentialdef/{creddef}:
    get:
      tags:
      - "credential"
      summary: "Retrieve a credential definition from the ledger"
      description: "Retrieve a credential definition from the ledger"
      operationId: "apiCredentialdefCreddefGET"
      parameters:
      - name: "creddef"
        in: "path"
        description: "A unique string value identifying this credential definition."
        required: true
        type: "string"
      - name: "wallet"
        in: "header"
        description: "(Optional) A unique id value (or \"default\") identifying a\
          \ wallet"
        required: false
        type: "string"
      responses:
        200:
          description: ""
      x-swagger-router-controller: "Credential"
    delete:
      tags:
      - "credential"
      summary: "Delete a credential definition"
      description: "Delete a credential definition"
      operationId: "apiCredentialdefCreddefDELETE"
      parameters:
      - name: "creddef"
        in: "path"
        description: "A unique string value identifying this credential definition."
        required: true
        type: "string"
      - name: "wallet"
        in: "header"
        description: "(Optional) A unique id value (or \"default\") identifying a\
          \ wallet"
        required: false
        type: "string"
      responses:
        204:
          description: ""
      x-swagger-router-controller: "Credential"
  /api/credentialoffer:
    post:
      tags:
      - "credential"
      summary: "Create a credential offer"
      description: "Create a credential offer"
      operationId: "apiCredentialofferPOST"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/credentialoffer_post"
      responses:
        201:
          description: ""
      x-swagger-router-controller: "Credential"
  /api/credentialrequest:
    post:
      tags:
      - "credential"
      summary: "Accept credential offer and create credential request"
      description: "Accept credential offer and create credential request"
      operationId: "apiCredentialrequestPOST"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/credentialrequest_post"
      responses:
        201:
          description: ""
      x-swagger-router-controller: "Credential"
  /api/credentialissue:
    post:
      tags:
      - "credential"
      summary: "Issue a credential"
      description: "Issue a credential"
      operationId: "apiCredentialissuePOST"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/credentialissue_post"
      responses:
        201:
          description: ""
      x-swagger-router-controller: "Credential"
  /api/credential/:
    get:
      tags:
      - "credential"
      summary: "List all credentials of specific wallet (UNTESTED ENDPOINT)"
      description: "List all credentials of specific wallet (UNTESTED ENDPOINT)"
      operationId: "apiCredentialGET"
      parameters:
      - name: "wallet"
        in: "header"
        description: "(Optional) A unique id value (or \"default\") identifying a\
          \ wallet"
        required: false
        type: "string"
      - name: "schema"
        in: "query"
        description: "(Optional) schemaId to filter credentials"
        required: false
        type: "string"
      - name: "schemaIssuerDid"
        in: "query"
        description: "(Optional) schemaIssuerDid to filter credentials"
        required: false
        type: "string"
      - name: "schemaName"
        in: "query"
        description: "(Optional) schemaName to filter credentials"
        required: false
        type: "string"
      - name: "schemaVersion"
        in: "query"
        description: "(Optional) schemaVersion to filter credentials"
        required: false
        type: "string"
      - name: "issuerDid"
        in: "query"
        description: "(Optional) issuerDid to filter credentials"
        required: false
        type: "string"
      - name: "credDefId"
        in: "query"
        description: "(Optional) credDefId to filter credentials"
        required: false
        type: "string"
      responses:
        200:
          description: ""
      x-swagger-router-controller: "Credential"
    post:
      tags:
      - "credential"
      summary: "Store a credential"
      description: "Store a credential"
      operationId: "apiCredentialPOST"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/credential_post"
      responses:
        201:
          description: ""
      x-swagger-router-controller: "Credential"
  /api/credential/{id}:
    get:
      tags:
      - "credential"
      summary: "Retrieve a credential"
      description: "Retrieve a credential"
      operationId: "apiCredentialIdGET"
      parameters:
      - name: "id"
        in: "path"
        description: "A unique string value identifying this credential."
        required: true
        type: "string"
      - name: "wallet"
        in: "header"
        description: "(Optional) A unique id value (or \"default\") identifying a\
          \ wallet"
        required: false
        type: "string"
      responses:
        200:
          description: ""
      x-swagger-router-controller: "Credential"
...
securityDefinitions:
  Bearer:
    description: "JWT Authorization header using the Bearer scheme.\nExample: 'Authorization:\
      \ Bearer {token}'\n\nPlease add Bearer in front of the JWT token:\n> Bearer\
      \ xxxxx.yyyyy.zzzzz\n"
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
  wallet:
    type: "string"
    description: "(Optional) A unique id value identifying a wallet"
...
  credential_post:
    type: "object"
    required:
    - "encryptedCredential"
    properties:
      wallet:
        $ref: "#/definitions/wallet"
      encryptedCredential:
        type: "string"
        description: "The authcrypted credential offer id"
  credentialissue_post:
    type: "object"
    required:
    - "encryptedCredentialRequest"
    properties:
      wallet:
        $ref: "#/definitions/wallet"
      encryptedCredentialRequest:
        type: "string"
        description: "The authcrypted Credential Request"
      values:
        type: "object"
        description: "Object containing attributes defined in schema as key-value\
          \ pairs (e.g. {\"attrName\":\"attrValue\", \"attrName1\":\"attrValue1\"\
          })"
        properties: {}
  credentialdef_post:
    type: "object"
    required:
    - "schemaId"
    properties:
      wallet:
        $ref: "#/definitions/wallet"
      schemaId:
        type: "string"
        description: "A unique string value identifying a schema"
      supportRevocation:
        type: "boolean"
        description: "(Optional) States if revocation should be supported for this\
          \ credential definiton"
  credentialoffer_post:
    type: "object"
    required:
    - "credDefId"
    - "recipientDid"
    properties:
      wallet:
        $ref: "#/definitions/wallet"
      recipientDid:
        type: "string"
        description: "DID for whom to create a credential offer"
      credDefId:
        type: "string"
        description: "Credential definition id"
  credentialrequest_post:
    type: "object"
    required:
    - "encryptedCredentialOffer"
    properties:
      wallet:
        $ref: "#/definitions/wallet"
      encryptedCredentialOffer:
        type: "string"
        description: "auth crypted credential offer"
  connectionoffer_post:
    type: "object"
    properties:
      wallet:
        $ref: "#/definitions/wallet"
      endpoint:
        type: "string"
        description: "(Optional) Response endpoint for connection response - my endpoint"
      role:
        type: "string"
        description: "(Optional) Role offered to the connection invitee"
        enum:
        - "NONE"
        - "TRUSTEE"
        - "STEWARD"
        - "TRUST_ANCHOR"
      meta:
        type: "object"
        description: "(Optional) Additional data to store in the offer (and later\
          \ in the pairwise), e.g. citizen id but not send with the offer"
        properties: {}
      data:
        type: "object"
        description: "(Optional) Additional data like name, logo, description, to\
          \ send with the connection offer"
        properties: {}
  connectionrequest_post:
    type: "object"
    properties:
      wallet:
        $ref: "#/definitions/wallet"
      endpoint:
        type: "string"
        description: "(Optional) Response endpoint for connection response - my endpoint"
      theirDid:
        type: "string"
        description: "(Optional) Their (endpoint) did, required if there is no connection\
          \ offer"
      theirVk:
        type: "string"
        description: "(Optional) Their (endpoint) did verification key, required if\
          \ there is no connection offer or verkey not on ledger."
      theirEndpoint:
        type: "string"
        description: "(Optional) Their endpoint address, required if no there is no\
          \ connection offer or address not on ledger."
      connectionOffer:
        type: "object"
        example:
          id: "09348295728957894752397589247589247589"
          type: "urn:sovrin:agent:message_type:sovrin.org/connection_offer"
          message:
            did: "endpointDid"
            verkey: "endpointVk"
            endpoint: "endpointAddress"
            nonce: "09348295728957894752397589247589247589"
        description: "(Optional) A connection offer object"
        properties: {}
  connectionresponse_post:
    type: "object"
    required:
    - "connectionRequestId"
    properties:
      wallet:
        $ref: "#/definitions/wallet"
      connectionRequestId:
        type: "string"
        example: "5b7ddce00000000000000000"
        description: "A unique string value identifying the connection request to\
          \ respond to"
...
  inline_response_200:
    properties:
      myDid:
        type: "string"
        description: "my pairwise did for this connection"
      theirDid:
        type: "string"
        description: "their pairwise did for this connection"
      acknowledged:
        type: "boolean"
        description: "boolean value indicating if connection is acknowledged"
    example:
      theirDid: "theirDid"
      acknowledged: true
      myDid: "myDid"
...